修改脚本

+++ deploy-nfd.sh
@@ -31,11 +31,26 @@ function cloneOrUpdate() {
     cd ${DEFAULT_DIR}
     name=$1
     url=$2
-    if [[ ${FORCE_UPDATE} -eq 1 ]]; then
-        rm -r ${name} ${name}.tar.gz
-    fi
-    if [[ ! -d ${name} ]];then
+    version=$3
+    if [[ ! -d ${name} ]];then          # 项目不存在则为其创建文件夹
         mkdir ${name}
+    else
+        if [[ ${FORCE_UPDATE} -eq 1 ]]; then        # 强制更新则不管是否版本号相同都更新
+            rm -r ${name} ${name}.tar.gz
+            mkdir ${name}
+        else
+            # 得到当前的版本号
+            curVersion=$(cat ${name}/VERSION)
+
+            # 版本号不一致则删除原来的，更新新的
+            if [[ ${curVersion} != ${version} ]]; then
+                echo "${name}版本更新： ${curVersion} -> ${version}"
+                rm -r ${name} ${name}.tar.gz
+                mkdir ${name}
+            else
+                echo "${name}已是最新版本，无需更新"
+            fi
+        fi
     fi
     curl -L ${url} > ${name}.tar.gz
     tar xf ${name}.tar.gz -C ${name} --strip 1
@@ -54,7 +69,7 @@ sudo apt-get install doxygen graphviz python-sphinx python-pip -y
 sudo pip install sphinxcontrib-doxylink sphinxcontrib-googleanalytics
 
 # install ndn-cxx
-cloneOrUpdate ndn-cxx https://github.com/named-data/ndn-cxx/archive/ndn-cxx-${NDN_CXX_VERSION}.tar.gz
+cloneOrUpdate ndn-cxx https://github.com/named-data/ndn-cxx/archive/ndn-cxx-${NDN_CXX_VERSION}.tar.gz ${NDN_CXX_VERSION}
 ./waf configure --enable-static
 ./waf
 sudo ./waf install
@@ -63,7 +78,7 @@ sudo ./waf install
 sudo apt-get install build-essential pkg-config libboost-all-dev \
                      libsqlite3-dev libssl-dev libpcap-dev -y
 sudo apt-get install doxygen graphviz python-sphinx -y
-cloneOrUpdate NFD https://github.com/named-data/NFD/archive/NFD-${NDN_NFD_VERSION}.tar.gz
+cloneOrUpdate NFD https://github.com/named-data/NFD/archive/NFD-${NDN_NFD_VERSION}.tar.gz ${NDN_NFD_VERSION}
 mkdir websocketpp
 curl -L https://github.com/zaphoyd/websocketpp/archive/${WEB_SOCKET_PP_VERSION}.tar.gz > websocket.tar.gz
 tar xf websocket.tar.gz -C websocketpp/ --strip 1
@@ -71,34 +86,35 @@ tar xf websocket.tar.gz -C websocketpp/ --strip 1
 ./waf
 sudo ./waf install
 #check nfd.conf exists?
-if [[ ! -f /usr/local/etc/ndn/nfd.conf ]];then
-    cd /usr/local/etc/ndn
-    sudo cp nfd.conf.sample nfd.conf
-fi
+#if [[ ! -f /usr/local/etc/ndn/nfd.conf ]];then
+#    cd /usr/local/etc/ndn
+sudo cp nfd.conf.sample nfd.conf
+#fi
 
 # install ChronoSync
-cloneOrUpdate ChronoSync https://github.com/named-data/ChronoSync/archive/${CHRONO_SYNC_VERSION}.tar.gz
+cloneOrUpdate ChronoSync https://github.com/named-data/ChronoSync/archive/${CHRONO_SYNC_VERSION}.tar.gz ${CHRONO_SYNC_VERSION}
 ./waf configure
 ./waf
 sudo ./waf install
 
 # install PSync
-cloneOrUpdate PSync https://github.com/named-data/PSync/archive/${PSYNC_VERSION}.tar.gz
+cloneOrUpdate PSync https://github.com/named-data/PSync/archive/${PSYNC_VERSION}.tar.gz ${PSYNC_VERSION}
 ./waf configure
 ./waf
 sudo ./waf install
 
 
 # install NLSR
-cloneOrUpdate NLSR https://github.com/named-data/NLSR/archive/NLSR-${NLSR_VERSION}.tar.gz
+cloneOrUpdate NLSR https://github.com/named-data/NLSR/archive/NLSR-${NLSR_VERSION}.tar.gz ${NLSR_VERSION}
 ./waf configure
 ./waf
 sudo ./waf install
 #check nlsr.conf exists?
-if [[ ! -f /usr/local/etc/ndn/nlsr.conf ]];then
-    cd /usr/local/etc/ndn
+#if [[ ! -f /usr/local/etc/ndn/nlsr.conf ]];then
+#    cd /usr/local/etc/ndn
     sudo cp nlsr.conf.sample nlsr.conf
-fi
+#fi
+
 #check /var/lib/nlsr exists?
 if [[ ! -d /var/lib/nlsr ]];then
     sudo mkdir /var/lib/nlsr
@@ -106,7 +122,7 @@ fi
 
 # install ndn-tools
 sudo apt-get install libpcap-dev -y
-cloneOrUpdate ndn-tools https://github.com/named-data/ndn-tools/archive/ndn-tools-${NDN_TOOLS_VERSION}.tar.gz
+cloneOrUpdate ndn-tools https://github.com/named-data/ndn-tools/archive/ndn-tools-${NDN_TOOLS_VERSION}.tar.gz ${NDN_TOOLS_VERSION}
 ./waf configure
 ./waf
 sudo ./waf install
